notifications/requirements.txt
firebase-admin
djangorestframework
apscheduler
python-decouple